feat: add update user profile use case and fix navigation for onboarding UI

+++ app/src/main/java/com/example/nutrition_assessment_system_android_app/domain/repository/UserRepository.kt
@@ -28,7 +28,7 @@ interface UserRepository {
     suspend fun updateUserProfile(
         gender: String,
         age: Int,
-        height: Int,
+        height: Float,
         weight: Float,
         activityLevel: String,
         goal: String
